moralis-python-scripts/requirements.txt
moralis
requests
mysql-connector-python